Moved session data into session namespace
Keep the core namespace cleaner and easier to navigate by moving all the session
information into a separate namespace.

Kept the session helper functions in the core namespace, as they manipulate the
app-state.

Included all the ideas and trial code in the session namespace, again to keep
the core clean.